Additional refactoring of DiscreteOrdinatesProblem. Remove duplicated validation and sweep-dispatch code. Split boundary-management and sweep/FLUDS/AngleSet/SweepChunk-construction code out of discrete_ordinates_problem.cc into separate files.
