+++ modules/linear_boltzmann_solvers/discrete_ordinates_problem/sweep_chunks/cbcd_sweep_chunk.cu
@@ -34,7 +34,7 @@ CBCDSweepChunk::CBCDSweepChunk(DiscreteOrdinatesProblem& problem, LBSGroupset& g
     angle_sets_.push_back(angle_set);
     fluds_list_.push_back(fluds);
     streams_list_.push_back(angle_set->GetStream());
-    gpu_kernel::Arguments<gpu_kernel::SweepType::CBC> args(
+    gpu_kernel::Arguments<SweepKind::CBC> args(
       problem_, groupset_, *angle_set, *fluds, surface_source_active_);
     kernel_args_list_.push_back(args);
     unsigned int stride_size =
@@ -63,14 +63,14 @@ CBCDSweepChunk::Sweep(const std::vector<std::uint32_t>& cell_local_ids, std::siz
   crb::Dim3 grid_size(grid_size_x, grid_size_y);
   auto* host_cell_local_ids_data = host_cell_local_ids.data();
 #if defined(__NVCC__) || defined(__HIPCC__)
-  gpu_kernel::SweepKernel<gpu_kernel::SweepType::CBC><<<grid_size, block_size, 0, stream>>>(
+  gpu_kernel::SweepKernel<SweepKind::CBC><<<grid_size, block_size, 0, stream>>>(
     args, host_cell_local_ids_data, num_ready_cells, device_saved_psi);
 #elif defined(SYCL_LANGUAGE_VERSION) && defined(__INTEL_LLVM_COMPILER)
   stream.synchronize();
   stream.parallel_for(sycl::nd_range<3>(grid_size * block_size, block_size),
                       [=](sycl::nd_item<3> work_index)
                       {
-                        gpu_kernel::SweepKernel<gpu_kernel::SweepType::CBC>(
+                        gpu_kernel::SweepKernel<SweepKind::CBC>(
                           args, host_cell_local_ids_data, num_ready_cells, device_saved_psi);
                       });
 #endif
