+++ modules/linear_boltzmann_solvers/discrete_ordinates_problem/sweep_chunks/gpu_kernel/solver.h
@@ -15,7 +15,7 @@ namespace opensn::gpu_kernel
 {
 
 /// Compute the sweep matrix from gradient, mass, and the source term
-template <std::size_t ndofs, SweepType t>
+template <std::size_t ndofs, SweepKind k>
 __CRB_DEVICE_FUNC__ void
 ComputeGMS(double* sweep_matrix,
            double* psi,
@@ -24,7 +24,7 @@ ComputeGMS(double* sweep_matrix,
            DirectionView& direction,
            const unsigned int& group_idx,
            const std::uint32_t& num_moments,
-           const Arguments<t>& args)
+           const Arguments<k>& args)
 {
   // get sigmaT
   double sigma_t = cell.total_xs[args.groupset_start + group_idx];
@@ -60,7 +60,7 @@ ComputeGMS(double* sweep_matrix,
 }
 
 /// Compute the sweep matrix from surface integral
-template <std::size_t ndofs, SweepType t>
+template <std::size_t ndofs, SweepKind k>
 __CRB_DEVICE_FUNC__ void
 ComputeSurfaceIntegral(double* sweep_matrix,
                        double* psi,
@@ -69,7 +69,7 @@ ComputeSurfaceIntegral(double* sweep_matrix,
                        const std::uint64_t* cell_edge_data,
                        const unsigned int& angle_group_idx,
                        const unsigned int& group_idx,
-                       const Arguments<t>& args)
+                       const Arguments<k>& args)
 {
   // loop over each face
   std::uint32_t face_node_counter = 0;
@@ -79,7 +79,7 @@ ComputeSurfaceIntegral(double* sweep_matrix,
     FaceView face;
     cell.GetFaceView(face, f);
     // determine if this face is incoming
-    NodeIndexType<t> idx(cell_edge_data[face_node_counter]);
+    NodeIndexType<k> idx(cell_edge_data[face_node_counter]);
     if (not idx.IsUndefined() and not idx.IsOutgoing())
     {
       double mu = direction.omega[0] * face.normal[0] + direction.omega[1] * face.normal[1] +
@@ -95,7 +95,7 @@ ComputeSurfaceIntegral(double* sweep_matrix,
           double mu_Nij = -mu * face.M_surf_data[fi * face.num_face_nodes + fj];
           Ai[j] += mu_Nij;
           double* upwind_psi;
-          if constexpr (t == SweepType::AAH)
+          if constexpr (k == SweepKind::AAH)
             upwind_psi =
               args.flud_data.GetIncomingFluxPointer(cell_edge_data[face_node_counter + fj],
                                                     angle_group_idx,
@@ -160,7 +160,7 @@ GaussianElimination(double* sweep_matrix, double* psi)
 }
 
 /// Record angular flux to downwind and compute outflow for boundary faces.
-template <SweepType t>
+template <SweepKind k>
 __CRB_DEVICE_FUNC__ void
 WritePsiToFludsAndOutflow(double* psi,
                           CellView& cell,
@@ -168,7 +168,7 @@ WritePsiToFludsAndOutflow(double* psi,
                           const std::uint64_t* cell_edge_data,
                           const unsigned int& angle_group_idx,
                           const unsigned int& group_idx,
-                          const Arguments<t>& args)
+                          const Arguments<k>& args)
 {
   // loop over each face
   std::uint32_t face_node_counter = 0;
@@ -178,7 +178,7 @@ WritePsiToFludsAndOutflow(double* psi,
     FaceView face;
     cell.GetFaceView(face, f);
     // determine if this face is outgoing
-    NodeIndexType<t> idx(cell_edge_data[face_node_counter]);
+    NodeIndexType<k> idx(cell_edge_data[face_node_counter]);
     if (not idx.IsUndefined() and idx.IsOutgoing())
     {
       // compute outflow for boundary faces
@@ -195,7 +195,7 @@ WritePsiToFludsAndOutflow(double* psi,
       }
       // skip for non reflecting boundary
       bool write_downwind_psi = true;
-      if constexpr (t == SweepType::AAH)
+      if constexpr (k == SweepKind::AAH)
       {
         if (idx.IsBoundary() && !idx.IsReflecting())
         {
@@ -208,10 +208,10 @@ WritePsiToFludsAndOutflow(double* psi,
         for (std::uint32_t fi = 0; fi < face.num_face_nodes; ++fi)
         {
           std::uint32_t i = face.cell_mapping_data[fi];
-          NodeIndexType<t> node_index(cell_edge_data[face_node_counter + fi]);
+          NodeIndexType<k> node_index(cell_edge_data[face_node_counter + fi]);
           // put copy psi to FLUDS
           double* downwind_psi;
-          if constexpr (t == SweepType::AAH)
+          if constexpr (k == SweepKind::AAH)
             downwind_psi = args.flud_data.GetOutgoingFluxPointer(
               node_index, angle_group_idx, args.boundary, args.boundary_offset);
           else
@@ -225,14 +225,14 @@ WritePsiToFludsAndOutflow(double* psi,
 }
 
 /// Compute the scalar flux.
-template <std::size_t ndofs, SweepType t>
+template <std::size_t ndofs, SweepKind k>
 __CRB_DEVICE_FUNC__ void
 ComputePhi(double* psi,
            CellView& cell,
            DirectionView& direction,
            const unsigned int& group_idx,
            const std::uint32_t& num_moments,
-           const Arguments<t>& args)
+           const Arguments<k>& args)
 {
   double* phi = args.phi + cell.phi_address + args.groupset_start + group_idx;
   _Pragma("unroll") for (std::uint32_t i = 0; i < ndofs; ++i)
@@ -263,9 +263,9 @@ SaveAngularFlux(const double* psi,
 }
 
 /// Template device function performing the sweep
-template <std::size_t ndofs, SweepType t>
+template <std::size_t ndofs, SweepKind k>
 __CRB_DEVICE_FUNC__ void
-Sweep(const Arguments<t>& args,
+Sweep(const Arguments<k>& args,
       CellView& cell,
       DirectionView& direction,
       const std::uint64_t* cell_edge_data,
@@ -277,16 +277,16 @@ Sweep(const Arguments<t>& args,
   // initialize buffer
   Buffer<ndofs> buffer;
   // prepare linear system to solve
-  ComputeGMS<ndofs, t>(
+  ComputeGMS<ndofs, k>(
     buffer.A(), buffer.b(), buffer.s(), cell, direction, group_idx, num_moments, args);
-  ComputeSurfaceIntegral<ndofs, t>(
+  ComputeSurfaceIntegral<ndofs, k>(
     buffer.A(), buffer.b(), cell, direction, cell_edge_data, angle_group_idx, group_idx, args);
   // solve for the angular flux
   GaussianElimination<ndofs>(buffer.A(), buffer.b());
   // save the result
-  WritePsiToFludsAndOutflow<t>(
+  WritePsiToFludsAndOutflow<k>(
     buffer.b(), cell, direction, cell_edge_data, angle_group_idx, group_idx, args);
-  ComputePhi<ndofs, t>(buffer.b(), cell, direction, group_idx, num_moments, args);
+  ComputePhi<ndofs, k>(buffer.b(), cell, direction, group_idx, num_moments, args);
   if (saved_psi != nullptr)
   {
     SaveAngularFlux<ndofs>(
